+++ xarray/test/test_backends.py
@@ -922,9 +922,14 @@ class GenericNetCDFDataTest(CFEncodedDataTest, Only32BitTypes, TestCase):
 
     def test_encoding_unlimited_dims(self):
         ds = Dataset({'x': ('y', np.arange(10.0))})
+        with self.roundtrip(ds,
+                            save_kwargs=dict(unlimited_dims=['y'])) as actual:
+            self.assertEqual(actual.encoding['unlimited_dims'], set('y'))
+            self.assertDatasetEqual(ds, actual)
         ds.encoding = {'unlimited_dims': ['y']}
         with self.roundtrip(ds) as actual:
             self.assertEqual(actual.encoding['unlimited_dims'], set('y'))
+            self.assertDatasetEqual(ds, actual)
 
 
 @requires_h5netcdf
@@ -983,6 +988,9 @@ class H5NetCDFDataTest(BaseNetCDF4Test, TestCase):
         ds.encoding = {'unlimited_dims': ['y']}
         with pytest.warns(UserWarning):
             ds.to_netcdf('foo-bar.nc', engine='h5netcdf')
+        with pytest.raises(NotImplementedError):
+            ds.to_netcdf('foo-bar.nc', engine='h5netcdf', unlimited_dims=['y'])
+
 
 @requires_dask
 @requires_scipy